moved UI independent code into separate library
