+++ HzdTextureLib/HzDException.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace HzdTextureExplorer
+namespace HzdTextureLib
 {
     public class HzDException : Exception
     {
