+++ HzdTextureLib/ImageData.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.IO;
 
-namespace HzdTextureExplorer
+namespace HzdTextureLib
 {
     public class ImageData
     {
